iconic_exam_maker/requirements.txt
PySide6~=6.7.0
PyMuPDF~=1.24.0
Pillow~=10.3.0
opencv-python~=4.9.0.80
numpy~=1.26.0
python-docx~=1.1.0
ultralytics~=8.2.0
